wifi: Eliminate some Visual Studio compiler warnings

+++ src/wifi/model/ht-operation.cc
@@ -332,7 +332,7 @@ HtOperation::GetSerializedSize () const
 uint8_t
 HtOperation::GetInformationSubset1 (void) const
 {
-  uint16_t val = 0;
+  uint8_t val = 0;
   val |= m_secondaryChannelOffset & 0x03;
   val |= (m_staChannelWidth & 0x01) << 2;
   val |= (m_rifsMode & 0x01) << 3;
@@ -368,7 +368,7 @@ HtOperation::SetInformationSubset2 (uint16_t ctrl)
   m_nonGfHtStasPresent = (ctrl >> 2) & 0x01;
   m_reservedInformationSubset2_1 = (ctrl >> 3) & 0x01;
   m_obssNonHtStasPresent = (ctrl >> 4) & 0x01;
-  m_reservedInformationSubset2_1 = (ctrl >> 5) & 0x07ff;
+  m_reservedInformationSubset2_1 = static_cast<uint8_t>((ctrl >> 5) & 0x07ff);
 }
 
 uint16_t
@@ -477,8 +477,8 @@ HtOperation::DeserializeInformationField (Buffer::Iterator start,
   Buffer::Iterator i = start;
   uint8_t primarychannel = i.ReadU8 ();
   uint8_t informationsubset1 = i.ReadU8 ();
-  uint8_t informationsubset2 = i.ReadU16 ();
-  uint8_t informationsubset3 = i.ReadU16 ();
+  uint16_t informationsubset2 = i.ReadU16 ();
+  uint16_t informationsubset3 = i.ReadU16 ();
   uint64_t mcsset1 = i.ReadLsbtohU64 ();
   uint64_t mcsset2 = i.ReadLsbtohU64 ();
   SetPrimaryChannel (primarychannel);
